+++ includes/styles/_index.php
@@ -3,21 +3,28 @@
  * @package    Additional Block Styles
  * @copyright  WebMan Design, Oliver Juhas
  *
- * @since  1.0.0
+ * @since    1.0.0
+ * @version  1.8.1
  */
 
 // Exit if accessed directly.
 defined( 'ABSPATH' ) || exit;
 
-include 'a11y.php';
-include 'alignment.php';
-include 'gallery.php';
-include 'image.php';
-include 'lists.php';
-include 'media-text.php';
-include 'others.php';
-include 'overlap.php';
-include 'quotes.php';
-include 'responsiveness.php';
-include 'separator.php';
-include 'spacing.php';
+$files = array(
+	'a11y',
+	'alignment',
+	'gallery',
+	'image',
+	'lists',
+	'media-text',
+	'others',
+	'overlap',
+	'quotes',
+	'responsiveness',
+	'separator',
+	'spacing',
+);
+
+foreach ( $files as $file ) {
+	include_once __DIR__ . '/' . $file . '.php';
+}